+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+`13 May 2012`
+
+ * Add DateTimeRange and DateRange column
+ * Add parentType search for cell render
+
 `8 May 2012`
 
  * Nothing private!